# Use fast-safe-stringify for perf and to support circular refs