--- a/package.json
+++ b/package.json
@@ -27,6 +27,7 @@
   "homepage": "https://github.com/winstonjs/logform#readme",
   "dependencies": {
     "colors": "^1.2.1",
+    "fast-safe-stringify": "^2.0.4",
     "fecha": "^2.3.3",
     "ms": "^2.1.1",
     "triple-beam": "^1.2.0"