While catching up with @davidmarkclements at JSConf I was reminding that `fast-safe-stringify` is a great piece of code we should use by default. This PR:

- [x] Uses `fast-safe-stringify` instead of `JSON.stringify`
- [x] Adds a test for circular JSON since **bonus** `fast-safe-stringify` _also_ supports circular JSON refs 🎉 
- [x] Renames `writeable` to `writable` within internal test code to be consistent with Node core naming conventions.

I tip my hat to @davidmarkclements here for doing all the heavy lifting:

**Before**
```
==========
OBJECT benchmark averages
BunyanObj average: 706.089ms
WinstonObj average: 519.757ms
BoleObj average: 291.980ms
LogLevelObject average: 556.841ms
PinoObj average: 279.025ms
PinoUnsafeObj average: 277.313ms
PinoExtremeObj average: 154.406ms
PinoUnsafeExtremeObj average: 151.695ms
==========
```

**After**
```
==========
OBJECT benchmark averages
BunyanObj average: 727.584ms
WinstonObj average: 446.947ms
BoleObj average: 309.041ms
LogLevelObject average: 577.650ms
PinoObj average: 282.819ms
PinoUnsafeObj average: 282.245ms
PinoExtremeObj average: 155.538ms
PinoUnsafeExtremeObj average: 148.020ms
==========
```

That's almost a 20% gain on that benchmark. 💯